Add curve points helper
