+++ addons/waterways_net/Data/UI/RiverEditMode.cs
@@ -1,4 +1,4 @@
-﻿namespace Waterways.UI.Data;
+﻿namespace Waterways.Data.UI;
 
 public enum RiverEditMode
 {
